Added Read-data functionality
