+++ src/app/components/markdown-pane/markdown-pane.component.html
@@ -4,7 +4,8 @@
     </div>
     <div class="markdown-page">
         <div class="markdown-content">
-            <textarea [(ngModel)]="textBoxData" class="content">
+            <textarea class="content">
+                {{ textBoxData }}
             </textarea>
         </div>
     </div>
