+++ src/app/components/markdown-pane/markdown-pane.component.scss
